update make data

+++ License_Plate_Localization/make_data.py
@@ -3,10 +3,11 @@ import shutil
 import Global_Var
 import xml.dom.minidom as mnd
 
+
 def DealXMLFile(filePath):
     domTree = mnd.parse(filePath)
-    rootNode = domTree.documentElement
     # 所有annotation
+    rootNode = domTree.documentElement
     # folder 元素
     folder = rootNode.getElementsByTagName("folder")[0]
     # phone 元素
@@ -16,8 +17,8 @@ def DealXMLFile(filePath):
     height = size.getElementsByTagName("height")[0]
     width = size.getElementsByTagName("width")[0]
     depth = size.getElementsByTagName("depth")[0]
-
     content = folder.childNodes[0].data + "/" + filename.childNodes[0].data + " "
+
     # object 元素
     objects = rootNode.getElementsByTagName("object")
     for object in objects:
@@ -38,7 +39,7 @@ def DealXMLFile(filePath):
 
 
 class CCPDNameParams():
-    def __init__(self,filename):
+    def __init__(self, filename):
         self.index_of_low = [i for i, j in enumerate(filename) if j in ["_"]]
         self.index_of_middle = [i for i, j in enumerate(filename) if j in ["-"]]
         self.index_of_and = [i for i, j in enumerate(filename) if j in ["&"]]
@@ -54,17 +55,17 @@ class CCPDNameParams():
         self.light = int(filename[self.index_of_middle[4] + 1: self.index_of_middle[-1]])
         self.blur = int(filename[self.index_of_middle[-1] + 1: self.index_of_point[0]])
 
-    def CCPDNameProcess(self):
+    def CCPDNameToLabelProcess(self):
         content = f"{self.xmin},{self.ymin},{self.xmax},{self.ymax},0 "
         return content
 
 
 def SelectFileFromCCPD():
     ccpdPath = "/Users/lanceren/Downloads/GD_Dataset/Raw_Data/2019/CCPD2019/"
-    targetFolder = "/Users/lanceren/Desktop/CCPD_Picts"
-    targetFolder_Normal = "/Users/lanceren/Desktop/CCPD_Picts/Normal"
-    targetFolder_SpecialCar = "/Users/lanceren/Desktop/CCPD_Picts/SpecialCar"
-    targetFolder_Weather = "/Users/lanceren/Desktop/CCPD_Picts/Weather"
+    targetFolder = "/Users/lanceren/Desktop/CCPD_Picts/"
+    targetFolder_Normal = "/Users/lanceren/Desktop/CCPD_Picts/Normal/"
+    targetFolder_SpecialCar = "/Users/lanceren/Desktop/CCPD_Picts/SpecialCar/"
+    targetFolder_Weather = "/Users/lanceren/Desktop/CCPD_Picts/Weather/"
 
     if os.path.exists(targetFolder): shutil.rmtree(targetFolder)
     os.mkdir(targetFolder)
@@ -86,8 +87,7 @@ def SelectFileFromCCPD():
 
             for filename in files:
                 totalCount += 1
-                fullFileName = rt +"/" + filename
-
+                fullFileName = rt + "/" + filename
                 ccpdName = CCPDNameParams(filename)
 
                 if ccpdName.horizon <= 0 and ccpdName.vertical <= 0 and ccpdName.light >= 100 and ccpdName.blur >= 100:
@@ -119,12 +119,10 @@ def CreateDotNames():
             f.close()
 
 
-def CreateLabelTxt(mode = "train"):
-
+def CreateLabelTxt(mode="train"):
     LabelTxtPath = Global_Var.projectPath + "/" + "License_Plate_Localization/" + "data/dataset/gd_detect_train.txt" if mode == "train" else \
         Global_Var.projectPath + "/" + "License_Plate_Localization/" + "data/dataset/gd_detect_test.txt"
-    if os.path.exists(LabelTxtPath):
-        os.remove(LabelTxtPath)
+    if os.path.exists(LabelTxtPath):os.remove(LabelTxtPath)
     with open(LabelTxtPath, "w+", encoding='utf-8') as f:
         labelData = []
         labelNum = 0
@@ -135,8 +133,7 @@ def CreateLabelTxt(mode = "train"):
             dirs[:] = [d for d in dirs if not d[0] == '.']
             for filename in files:
                 fullFileName = rt + filename
-                content = Global_Var.projectPath + "/License_Plate_Localization/dataset/" + DealXMLFile(
-                    fullFileName) + os.linesep
+                content = Global_Var.projectPath + "/License_Plate_Localization/data/dataset/" + DealXMLFile(fullFileName) + os.linesep
                 if mode == "train":
                     labelData.append(content)
                 else:
@@ -151,7 +148,7 @@ def CreateLabelTxt(mode = "train"):
             for filename in files:
                 fullFileName = rt + filename
                 ccpdName = CCPDNameParams(filename)
-                content = fullFileName + " " + ccpdName.CCPDNameProcess() + os.linesep
+                content = fullFileName + " " + ccpdName.CCPDNameToLabelProcess() + os.linesep
                 if mode == "train":
                     labelData.append(content)
                 else:
@@ -166,6 +163,3 @@ def CreateLabelTxt(mode = "train"):
 CreateDotNames()
 CreateLabelTxt("train")
 CreateLabelTxt("test")
-
-
-
